rename: calculateVwBasedOnFigma -> generateStyleBasedOnFigma #91
