+++ frontend/src/utils/generateStyleBasedOnFigma.ts
